feat(route/1lou): add keyword search route
1lou.me exposes a GET keyword search at search.htm?keyword=<kw> whose
results are the same li.media.thread.tap markup as the forum listings, so
the existing listing+detail parse applies unchanged. Extract that parse
into util.ts (fetchThreads) and reuse it from both the generic route and a
new /search/:keyword route (keyword as a path param).

Verified: /1lou/search/奥本海默 returns 20 threads with .torrent enclosures.
