+++ lib/routes/1lou/index.ts
@@ -1,12 +1,6 @@
-import { load } from 'cheerio';
-
 import type { Route } from '@/types';
-import cache from '@/utils/cache';
-import got from '@/utils/got';
-import { parseDate } from '@/utils/parse-date';
-import timezone from '@/utils/timezone';
 
-const rootUrl = 'https://www.1lou.me';
+import { fetchThreads, rootUrl } from './util';
 
 export const handler = async (ctx) => {
     const { params } = ctx.req.param();
@@ -18,78 +12,7 @@ export const handler = async (ctx) => {
 
     const currentUrl = new URL(`${params && params.endsWith('.htm') ? params : `${params}.htm`}${queryString ? `?${queryString}` : ''}`, rootUrl).href;
 
-    const { data: response } = await got(currentUrl);
-
-    const $ = load(response);
-
-    const language = $('html').prop('lang');
-
-    let items = $('li.media.thread.tap:not(li.hidden-sm)')
-        .slice(0, limit)
-        .toArray()
-        .map((item) => {
-            item = $(item);
-
-            const subjectEl = item.find('div.subject').children('a').first();
-
-            return {
-                title: subjectEl.text(),
-                pubDate: timezone(parseDate(item.find('span.date').text()), +8),
-                link: new URL(subjectEl.prop('href'), rootUrl).href,
-                category: [
-                    item.find('a.text-secondary').text().replaceAll('[]', ''),
-                    ...item
-                        .find('a.badge')
-                        .toArray()
-                        .map((c) => $(c).text()),
-                ].filter(Boolean),
-                author: item.find('a.username').text(),
-                language,
-            };
-        });
-
-    items = await Promise.all(
-        items.map((item) =>
-            cache.tryGet(item.link, async () => {
-                const { data: detailResponse } = await got(item.link);
-
-                const $$ = load(detailResponse);
-
-                const title = $$('h4.break-all').contents().last().text();
-
-                if (title) {
-                    const description = $$('div.message.break-all').html();
-                    const image = new URL($$('img').first().prop('src'), rootUrl).href;
-
-                    item.title = title;
-                    item.description = description;
-                    item.pubDate = timezone(parseDate($$('span.date').text()), +8);
-                    item.category = $$('a.badge')
-                        .toArray()
-                        .map((c) => $$(c).text());
-                    item.content = {
-                        html: description,
-                        text: $$('div.message.break-all').text(),
-                    };
-                    item.image = image;
-                    item.banner = image;
-                    item.language = language;
-
-                    const torrents = $$('ul.attachlist li a');
-
-                    if (torrents.length > 0) {
-                        const torrent = torrents.first();
-
-                        item.enclosure_url = new URL(torrent.prop('href'), rootUrl).href;
-                        item.enclosure_type = 'application/x-bittorrent';
-                        item.enclosure_title = torrent.text();
-                    }
-                }
-
-                return item;
-            })
-        )
-    );
+    const { $, items, language } = await fetchThreads(currentUrl, limit);
 
     const author = 'BT 之家 1LOU 站';
     const image = new URL($('img.logo-2').prop('src'), rootUrl).href;
@@ -121,7 +44,7 @@ export const route: Route = {
 
 若订阅 [最新发帖电视剧](https://www.1lou.me/forum-2-1.htm?orderby=tid\\&digest=0)，网址为 \`https://www.1lou.me/forum-2-1.htm?orderby=tid&digest=0\`。截取 \`https://www.1lou.me/\` 到末尾 \`.htm\` 的部分 \`forum-2-1\` 作为参数，并补充 \`orderby\`，此时路由为 [\`/1lou/forum-2-1?orderby=tid\`](https://rsshub.app/1lou/forum-2-1?orderby=tid)。
 
-若订阅 [搜素繁花主题贴](https://www.1lou.me/search-_E7_B9_81_E8_8A_B1-1.htm)，网址为 \`https://www.1lou.me/search-_E7_B9_81_E8_8A_B1-1.htm\`。截取 \`https://www.1lou.me/\` 到末尾 \`.htm\` 的部分 \`search-_E7_B9_81_E8_8A_B1-1\` 作为参数，此时路由为 [\`/1lou/search-_E7_B9_81_E8_8A_B1-1\`](https://rsshub.app/1lou/search-_E7_B9_81_E8_8A_B1-1)。
+若搜索关键词，请使用 [\`/1lou/search/:keyword\`](https://rsshub.app/1lou/search/奥本海默) 路由。
 :::`,
     categories: ['multimedia'],
